[Refactor/PYOKEMON-15] 예매 사이드바 컴포넌트화
