+++ src/routes.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter } from "react-router"
 
 import AdminProtectedRoute from "./components/admin-protected-route"
 import ProtectedRoute from "./components/protected-route"
-import EventDetailPage from "./pages/event-detail-page"
+import EventDetailPage from "./pages/event-detail/event-detail-page"
 import EventListPage from "./pages/event-list-page"
 import HomePage from "./pages/home/home-page"
 import MainEmptyLayout from "./pages/main-empty-layout"
